graphics

+++ src/com/hazam/widget/DownloadTask.java
@@ -24,40 +24,60 @@ import android.util.Log;
 import com.hazam.handy.fs.FileUtils;
 import com.hazam.handy.net.BetterHttpClient;
 
-public class DownloadTask extends AsyncTask<Uri, Long, Uri> implements FileUtils.Tick {
+public class DownloadTask extends AsyncTask<Void, Long, Uri> implements FileUtils.Tick {
 
 	private final Context ctx;
+	private final File cacheDir;
 	private final ConnectivityManager connectivityManager;
 	private final BetterHttpClient httpClient;
+
 	private long clength;
 	private Throwable error = null;
+	private Uri targetUri;
+	private File targetFile;
+	private DownloadListener listener;
 
-	public DownloadTask(final Context _ctx) {
+	public DownloadTask(final Context _ctx, Uri targetUri) {
 		this.ctx = _ctx.getApplicationContext();
 		this.connectivityManager = (ConnectivityManager) ctx.getSystemService(Activity.CONNECTIVITY_SERVICE);
 		this.httpClient = new BetterHttpClient(null, false);
+		this.cacheDir = new File(ctx.getCacheDir(), TAG);
+		this.targetUri = targetUri;
+	}
+
+	public static interface DownloadListener {
+		
+		public void onResourceUpdate(Uri uri);
+
+		public void onError(Throwable e);
 	}
 
 	@Override
 	protected void onPreExecute() {
+		String computed = targetUri.getLastPathSegment();
+		targetFile = new File(cacheDir, computed);
+		if (targetFile.exists() && listener != null) {
+			listener.onResourceUpdate(Uri.parse(targetFile.getAbsolutePath()));
+		}
 	}
 
 	@Override
 	protected void onPostExecute(final Uri result) {
 		if (error != null) {
-			//exception
+			// exception
+			if (listener != null) {
+				listener.onError(error);
+			}
 		} else if (result != null) {
-			//handle succesfull download or found in cache
+			// handle succesfull download or found in cache
+			if (listener != null) {
+				listener.onResourceUpdate(result);
+			}
 		} else {
-			//no exception, no uri: we don't have connection and no cached res is available
+			// no exception, no uri: we don't have connection and no cached res is available
 		}
 	}
-
-	private final File composeTargetFile(final Uri src) {
-		final String computed = src.getLastPathSegment();
-		return new File(ctx.getCacheDir(), computed);
-	}
-
+	
 	private void augmentWithIfModifiedSince(final File targetFile, final AbstractHttpMessage mess) {
 		if (targetFile.exists()) {
 			final String lastModifiedFormatted = DateUtils.formatDate(new Date(targetFile.lastModified()));
@@ -66,15 +86,17 @@ public class DownloadTask extends AsyncTask<Uri, Long, Uri> implements FileUtils
 	}
 
 	@Override
-	protected Uri doInBackground(final Uri... params) {
+	protected Uri doInBackground(final Void... params) {
+		if (!cacheDir.exists()) {
+			cacheDir.mkdirs();
+		}
 		error = null;
-		final Uri target = params[0];
-		final File targetFile = composeTargetFile(target);
 		if (weAreOnline()) {
-			final HttpGet getFile = new HttpGet(target.toString());
+			final HttpGet getFile = new HttpGet(targetUri.toString());
 			augmentWithIfModifiedSince(targetFile, getFile);
+			File tempFile = null;
 			try {
-				final File tempFile = File.createTempFile(TAG, ".tmp", ctx.getCacheDir());
+				tempFile = File.createTempFile(TAG, ".tmp", cacheDir);
 				final HttpResponse resp = httpClient.execute(getFile);
 				final HttpEntity ent = resp.getEntity();
 				final int statusCode = resp.getStatusLine().getStatusCode();
@@ -85,14 +107,19 @@ public class DownloadTask extends AsyncTask<Uri, Long, Uri> implements FileUtils
 				case HttpStatus.SC_NOT_MODIFIED:
 					break;
 				default:
-					throw new RuntimeException("Http Status Code: "+statusCode);
+					throw new RuntimeException("Http Status Code: " + statusCode);
 				}
 
 				if (ent != null) {
 					ent.consumeContent();
 				}
 			} catch (Throwable e) {
-				error = new RuntimeException(TAG+": error", e);
+				error = new RuntimeException(TAG + ": error", e);
+				e.printStackTrace();
+			} finally {
+				if (tempFile != null && tempFile.exists()) {
+					tempFile.delete();
+				}
 			}
 		}
 
@@ -103,15 +130,19 @@ public class DownloadTask extends AsyncTask<Uri, Long, Uri> implements FileUtils
 		}
 	}
 
-	private void handleEntity(final HttpEntity ent, final File targetFile, final File tempFile) throws IllegalStateException, IOException {
+	private void handleEntity(final HttpEntity ent, final File targetFile, final File tempFile)
+			throws IllegalStateException, IOException {
 		clength = ent.getContentLength();
 		InputStream in = ent.getContent();
-		FileOutputStream f = new FileOutputStream(tempFile);
-		FileUtils.decantStreams(in, f, this);
-		f.close();
-		targetFile.delete();
+		FileOutputStream out = new FileOutputStream(tempFile);
+		FileUtils.copy(in, out, this);
+		out.close();
+
+		// if everything went fine you want to switch temp and target file, guarded
+		if (targetFile.exists()) {
+			targetFile.delete();
+		}
 		FileUtils.copyFile(tempFile, targetFile);
-		tempFile.delete();
 	}
 
 	private static final String TAG = "DownloadTask";
@@ -133,5 +164,9 @@ public class DownloadTask extends AsyncTask<Uri, Long, Uri> implements FileUtils
 	private final boolean weAreOnline() {
 		final NetworkInfo ni = connectivityManager.getActiveNetworkInfo();
 		return ni != null && ni.isConnected();
+	}
+	
+	public void setListener(DownloadListener listener) {
+		this.listener = listener;
 	}
 }
